% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{arithmetic.mean}
\alias{arithmetic.mean}
\alias{geometric.mean}
\alias{harmonic.mean}
\alias{quadratic.mean}
\title{You're a \code{mean} one, Mr. Grinch}
\usage{
arithmetic.mean(x, ...)

geometric.mean(x, ..., zero.propagate = FALSE)

harmonic.mean(x, ...)

quadratic.mean(x, ...)
}
\arguments{
\item{x}{An R object. Supports all type that \code{mean()} supports.}

\item{...}{Additional arguments such as \code{trim} or \code{na.rm}}

\item{zero.propagate}{Logical, default \code{FALSE}. Should 0's be dropped
or propgated?}
}
\value{
The arithmetic mean
}
\description{
You're a \code{mean} one, Mr. Grinch
}
